pyproject.toml
[tool.poetry]
name = "senior-pomidor"
version = "0.1.0"
description = "learn Django"
authors = ["anton <sparco114@ya.ru>"]
readme = "README.md"
packages = [{include = "senior_pomidor"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
